Split functions
Split the functions into different files for easier testing
