+++ Leaderboard.py
@@ -1,0 +1,25 @@
+'''
+6. Leaderboard Function =========================================================================================================================================================
+
+'''
+def loadLeaderBoard():
+	print("To load leaderboard data into CSV")
+	#Step 1: Check exsisting leaderboard CSV file
+	#Step 2.1: Create new leaderboard CSV file
+	#Step 2.2: Load CSV data into list
+	#Step 3: return data list
+
+def createLeaderboardFile():
+	print("If there is no leaderboard CSV file, create leaderboard file")
+
+def saveLeaderboard():
+	print("To save leaderboard scores into a csv file")
+	#Step 1: Load leaderboard CSV file
+	leaderBoardData = loadLeaderBoard('Leaderboard.csv');
+	#Step 3: Add data into list
+	#Step 4: Save list into CSV file (Either overwrite or add on)
+
+def printLeaderboard():
+	print("print leaderboard")
+	#Step 1: Load leaderboard CSV file
+	#Step 2: Print leaderboard
